+++ src/Game/Werewolf/Command/Protector.hs
@@ -1,15 +1,15 @@
 {-|
-Module      : Game.Werewolf.Command.Defender
-Description : Defender commands.
+Module      : Game.Werewolf.Command.Protector
+Description : Protector commands.
 
 Copyright   : (c) Henry J. Wylde, 2016
 License     : BSD3
 Maintainer  : public@hjwylde.com
 
-Defender commands.
+Protector commands.
 -}
 
-module Game.Werewolf.Command.Defender (
+module Game.Werewolf.Command.Protector (
     -- * Commands
     protectCommand,
 ) where
@@ -28,8 +28,8 @@ import Game.Werewolf.Util
 protectCommand :: Text -> Text -> Command
 protectCommand callerName targetName = Command $ do
     validatePlayer callerName callerName
-    unlessM (isPlayerDefender callerName)                           $ throwError [playerCannotDoThatMessage callerName]
-    unlessM isDefendersTurn                                         $ throwError [playerCannotDoThatRightNowMessage callerName]
+    unlessM (isPlayerProtector callerName)                          $ throwError [playerCannotDoThatMessage callerName]
+    unlessM isProtectorsTurn                                        $ throwError [playerCannotDoThatRightNowMessage callerName]
     validatePlayer callerName targetName
     whenM (has (priorProtect . traverse . only targetName) <$> get) $ throwError [playerCannotProtectSamePlayerTwiceInARowMessage callerName]