+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ### Upcoming
 
+*Minor*
+
+* Renamed the Defender to Protector. ([#132](https://github.com/hjwylde/werewolf/issues/132))
+
 ### v0.5.0.0
 
 *Major*